requirements.txt
pydantic
fastapi
python-dateutil
requests
psycopg2
uvicorn
sqlalchemy